Move authentication at the root
