+++ src/communications/conference.ts
@@ -1,7 +1,8 @@
 import { sendDelete, sendPost, sendPut } from '../internal/httpHelpers';
+import { COMMS_HOSTNAME } from './internal/urls';
 import { CreateConferenceOptions, Conference, UserTokens } from './types/conference';
 import { SpatialEnvironment, SpatialListener, SpatialUsers } from './types/spatialAudio';
-import JwtToken from './types/jwtToken';
+import JwtToken from '../types/jwtToken';
 import Participant from './types/participant';
 import { RTCPMode } from './types/rtcpMode';
 
@@ -46,7 +47,7 @@ export const createConference = async (accessToken: JwtToken, options: CreateCon
     }
 
     const requestOptions = {
-        hostname: 'comms.api.dolby.io',
+        hostname: COMMS_HOSTNAME,
         path: '/v2/conferences/create',
         headers: {
             Accept: 'application/json',
@@ -86,7 +87,7 @@ export const invite = async (accessToken: JwtToken, conferenceId: string, partic
     });
 
     const options = {
-        hostname: 'comms.api.dolby.io',
+        hostname: COMMS_HOSTNAME,
         path: `/v2/conferences/${conferenceId}/invite`,
         headers: {
             Accept: 'application/json',
@@ -115,7 +116,7 @@ export const kick = async (accessToken: JwtToken, conferenceId: string, external
     });
 
     const options = {
-        hostname: 'comms.api.dolby.io',
+        hostname: COMMS_HOSTNAME,
         path: `/v2/conferences/${conferenceId}/kick`,
         headers: {
             Accept: 'application/json',
@@ -156,7 +157,7 @@ export const sendMessage = async (
     }
 
     const options = {
-        hostname: 'comms.api.dolby.io',
+        hostname: COMMS_HOSTNAME,
         path: `/v2/conferences/${conferenceId}/message`,
         headers: {
             Accept: 'application/json',
@@ -194,7 +195,7 @@ export const setSpatialListenersAudio = async (
     });
 
     const options = {
-        hostname: 'comms.api.dolby.io',
+        hostname: COMMS_HOSTNAME,
         path: `/v2/conferences/${conferenceId}/spatial-listeners-audio`,
         headers: {
             Accept: 'application/json',
@@ -233,7 +234,7 @@ export const updatePermissions = async (accessToken: JwtToken, conferenceId: str
     });
 
     const options = {
-        hostname: 'comms.api.dolby.io',
+        hostname: COMMS_HOSTNAME,
         path: `/v2/conferences/${conferenceId}/invite`,
         headers: {
             Accept: 'application/json',
@@ -257,7 +258,7 @@ export const updatePermissions = async (accessToken: JwtToken, conferenceId: str
  */
 export const terminate = async (accessToken: JwtToken, conferenceId: string): Promise<void> => {
     const options = {
-        hostname: 'comms.api.dolby.io',
+        hostname: COMMS_HOSTNAME,
         path: `/v2/conferences/${conferenceId}`,
         headers: {
             Accept: 'application/json',
