+++ src/communications/internal/urls.ts
@@ -1,0 +1,2 @@
+export const COMMS_HOSTNAME = 'comms.api.dolby.io';
+export const SESSION_HOSTNAME = 'session.voxeet.com';
