+++ src/internal/urls.ts
@@ -1,0 +1,1 @@
+export const AUTH_HOSTNAME = 'api.dolby.io';
