+++ src/media/analyzeSpeech.ts
@@ -1,5 +1,6 @@
 import { sendPost, sendGet, AuthRequestOptions } from './internal/httpHelpers';
-import JwtToken from './types/jwtToken';
+import { API_HOSTNAME } from './internal/urls';
+import JwtToken from '../types/jwtToken';
 import { AnalyzeSpeechJob } from './types/analyzeSpeech';
 
 /**
@@ -23,7 +24,7 @@ import { AnalyzeSpeechJob } from './types/analyzeSpeech';
  */
 export const start = async (accessToken: JwtToken, jobContent: string): Promise<string | null> => {
     const requestOptions: AuthRequestOptions = {
-        hostname: 'api.dolby.com',
+        hostname: API_HOSTNAME,
         path: '/media/analyze/speech',
         headers: {
             Accept: 'application/json',
@@ -57,7 +58,7 @@ export const start = async (accessToken: JwtToken, jobContent: string): Promise<
  */
 export const getResults = async (accessToken: JwtToken, jobId: string): Promise<AnalyzeSpeechJob> => {
     const requestOptions: AuthRequestOptions = {
-        hostname: 'api.dolby.com',
+        hostname: API_HOSTNAME,
         path: '/media/analyze/speech',
         params: {
             job_id: jobId,
