+++ src/media/diagnose.ts
@@ -1,5 +1,6 @@
 import { sendPost, sendGet, AuthRequestOptions } from './internal/httpHelpers';
-import JwtToken from './types/jwtToken';
+import { API_HOSTNAME } from './internal/urls';
+import JwtToken from '../types/jwtToken';
 import { DiagnoseJob } from './types/diagnose';
 
 /**
@@ -23,7 +24,7 @@ import { DiagnoseJob } from './types/diagnose';
  */
 export const start = async (accessToken: JwtToken, jobContent: string): Promise<string | null> => {
     const requestOptions: AuthRequestOptions = {
-        hostname: 'api.dolby.com',
+        hostname: API_HOSTNAME,
         path: '/media/diagnose',
         headers: {
             Accept: 'application/json',
@@ -59,7 +60,7 @@ export const start = async (accessToken: JwtToken, jobContent: string): Promise<
  */
 export const getResults = async (accessToken: JwtToken, jobId: string): Promise<DiagnoseJob> => {
     const requestOptions: AuthRequestOptions = {
-        hostname: 'api.dolby.com',
+        hostname: API_HOSTNAME,
         path: '/media/diagnose',
         params: {
             job_id: jobId,
