+++ src/media/enhance.ts
@@ -1,5 +1,6 @@
 import { sendPost, sendGet, AuthRequestOptions } from './internal/httpHelpers';
-import JwtToken from './types/jwtToken';
+import { API_HOSTNAME } from './internal/urls';
+import JwtToken from '../types/jwtToken';
 import { EnhanceJob } from './types/enhance';
 
 /**
@@ -20,7 +21,7 @@ import { EnhanceJob } from './types/enhance';
  */
 export const start = async (accessToken: JwtToken, jobContent: string): Promise<string | null> => {
     const requestOptions: AuthRequestOptions = {
-        hostname: 'api.dolby.com',
+        hostname: API_HOSTNAME,
         path: '/media/enhance',
         headers: {
             Accept: 'application/json',
@@ -54,7 +55,7 @@ export const start = async (accessToken: JwtToken, jobContent: string): Promise<
  */
 export const getResults = async (accessToken: JwtToken, jobId: string): Promise<EnhanceJob> => {
     const requestOptions: AuthRequestOptions = {
-        hostname: 'api.dolby.com',
+        hostname: API_HOSTNAME,
         path: '/media/enhance',
         params: {
             job_id: jobId,
