+++ src/media/index.ts
@@ -6,7 +6,6 @@ import * as io from './io';
 import * as jobs from './jobs';
 import * as mastering from './mastering';
 import * as transcode from './transcode';
-import * as platform from './platform';
 import * as webhooks from './webhooks';
 
-export { analyze, analyzeSpeech, diagnose, enhance, io, jobs, mastering, transcode, platform, webhooks };
+export { analyze, analyzeSpeech, diagnose, enhance, io, jobs, mastering, transcode, webhooks };
