+++ src/media/internal/urls.ts
@@ -1,0 +1,1 @@
+export const API_HOSTNAME = 'api.dolby.com';
