+++ src/media/jobs.ts
@@ -1,5 +1,6 @@
 import { sendGet, sendPost, AuthRequestOptions } from './internal/httpHelpers';
-import JwtToken from './types/jwtToken';
+import { API_HOSTNAME } from './internal/urls';
+import JwtToken from '../types/jwtToken';
 import { ListJobsOptions, ListAllJobsOptions, JobsResponse, Job } from './types/jobs';
 
 /**
@@ -29,7 +30,7 @@ export const list = async (accessToken: JwtToken, options: ListJobsOptions): Pro
     }
 
     const requestOptions: AuthRequestOptions = {
-        hostname: 'api.dolby.com',
+        hostname: API_HOSTNAME,
         path: '/media/jobs',
         params,
         headers: {},
@@ -64,7 +65,7 @@ export const listAll = async (accessToken: JwtToken, options: ListAllJobsOptions
     }
 
     const requestOptions: AuthRequestOptions = {
-        hostname: 'api.dolby.com',
+        hostname: API_HOSTNAME,
         path: '/media/jobs',
         headers: {},
         accessToken,
@@ -106,7 +107,7 @@ export const cancel = async (accessToken: JwtToken, jobId: string): Promise<void
     };
 
     const requestOptions: AuthRequestOptions = {
-        hostname: 'api.dolby.com',
+        hostname: API_HOSTNAME,
         path: '/media/jobs/cancel',
         headers: {
             Accept: 'application/json',
