+++ src/types/jwtToken.ts
@@ -2,5 +2,4 @@ export default interface JwtToken {
     access_token: string;
     expires_in: number;
     token_type: string;
-    status: string;
 }
